+++ backend/src/test/java/hanglog/trip/presentation/ItemControllerTest.java
@@ -8,6 +8,9 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import hanglog.trip.presentation.dto.request.ExpenseRequest;
+import hanglog.trip.presentation.dto.request.ItemRequest;
+import hanglog.trip.presentation.dto.request.PlaceRequest;
 import hanglog.trip.service.ItemService;
 import java.math.BigDecimal;
 import org.junit.jupiter.api.DisplayName;
